fix(backend): fix PR

+++ backend/internal/storage/table.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 // Process model struct for database table
 type Process struct {
 	gorm.Model
-	ProcessID    uint   `gorm:"unique;not null"`
-	ProcessKey   uint   `gorm:"unique;not null"`
+	ProcessID    int64  `gorm:"unique;not null"`
+	ProcessKey   int64  `gorm:"unique;not null"`
 	BpmnResource string `gorm:"not null"`
 }
